Moved `CurrentPlayer` from Features to Utils, as it's a support class used in Features as well as in Patches and not necessarily a Feature itself (khanshoaib3/stardew-access#112).
