+++ stardew-access/Features/Radar.cs
@@ -1,6 +1,7 @@
 using Microsoft.Xna.Framework;
 using StardewValley;
 using StardewValley.Objects;
+using stardew_access.Utils;
 
 namespace stardew_access.Features
 {
