+++ stardew-access/Features/TileViewer.cs
@@ -2,6 +2,7 @@
 using xTile;
 using StardewValley;
 using StardewValley.Menus;
+using stardew_access.Utils;
 
 namespace stardew_access.Features
 {
