+++ stardew-access/Features/Warnings.cs
@@ -1,3 +1,5 @@
+using stardew_access.Utils;
+
 namespace stardew_access.Features
 {
     /// <summary>
