+++ stardew-access/ModEntry.cs
@@ -1,10 +1,11 @@
-﻿using stardew_access.Features;
-using StardewModdingAPI;
+﻿using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
 using HarmonyLib;
+using stardew_access.Features;
 using stardew_access.Patches;
 using stardew_access.ScreenReader;
+using stardew_access.Utils;
 using Microsoft.Xna.Framework;
 
 namespace stardew_access
